requirements.txt
pandas
numpy
scikit-learn
yfinance
requests
beautifulsoup4
newspaper3k
tensorflow
torch
coinbase-advanced-py
python-dotenv
pyjwt